[1865] Refactor User access to be Editors
