+++ bin/mcb
@@ -22,14 +22,17 @@ require 'mcb'
 require 'mcb/azure'
 require 'mcb/config'
 require 'mcb/course_show'
+
 require 'mcb/cli/base_cli'
 require 'mcb/cli/course_cli'
 require 'mcb/cli/provider_cli'
+
 require 'mcb/editor/base'
 require 'mcb/editor/courses_editor'
 require 'mcb/editor/provider_editor'
-require 'mcb/grant_access_wizard'
-require 'mcb/revoke_access_wizard'
+require 'mcb/editor/grant_access_wizard'
+require 'mcb/editor/revoke_access_wizard'
+
 require 'mcb/render'
 require 'mcb/render/active_record'
 require 'mcb/render/apiv1'
